Capa de seguridad en registro y edicion de productos

+++ backend/controllers/usuario.controller.js
@@ -27,28 +27,35 @@ const getUsuarios = async(req = request, res = response) => {
 
 const crearUsuario = async(req = request, res = response) => {
     try {
-        const { email, password } = req.body;
-        const existeEmail = await UsuarioDb.findOne({ email });
+        if (req.rol == 'Administrador') {
+            const { email, password } = req.body;
+            const existeEmail = await UsuarioDb.findOne({ email });
 
-        if (existeEmail) {
-            return res.status(400).json({
+            if (existeEmail) {
+                return res.status(400).json({
+                    ok: false,
+                    msg: "El correo ya ha sido registrado"
+                });
+            }
+
+            const usuario = new UsuarioDb(req.body);
+
+            //Encriptacin de contraseña
+            const salt = bcrypt.genSaltSync();
+            usuario.password = bcrypt.hashSync(password, salt);
+
+            await usuario.save();
+
+            return res.status(200).send({
+                ok: true,
+                usuario
+            });
+        } else {
+            return res.status(400).send({
                 ok: false,
-                msg: "El correo ya ha sido registrado"
+                msg: 'El usuario no tiene permisos para acceder a esta información',
             });
         }
-
-        const usuario = new UsuarioDb(req.body);
-
-        //Encriptacin de contraseña
-        const salt = bcrypt.genSaltSync();
-        usuario.password = bcrypt.hashSync(password, salt);
-
-        await usuario.save();
-
-        return res.status(200).send({
-            ok: true,
-            usuario
-        });
 
     } catch (error) {
         return res.status(500).send({
@@ -62,36 +69,40 @@ const actualizarUsurario = async(req = request, res = response) => {
 
     const uid = req.params.id;
     try {
+        if (req.rol == 'Administrador') {
+            const usuario = await UsuarioDb.findById(uid);
 
-        const usuario = await UsuarioDb.findById(uid);
-
-        if (!usuario) {
-            return res.status(404).send({
-                ok: false,
-                msg: 'No existe un usuario por ese id'
-            });
-        }
-
-        const { password, email, ...campos } = req.body;
-
-        if (usuario.email !== email) {
-            const existeEmail = await UsuarioDb.findOne({ email });
-            if (existeEmail) {
-                return res.status(400).send({
+            if (!usuario) {
+                return res.status(404).send({
                     ok: false,
-                    msg: "Ya existe un usuario con ese email"
+                    msg: 'No existe un usuario por ese id'
                 });
             }
+
+            const { password, email, ...campos } = req.body;
+
+            if (usuario.email !== email) {
+                const existeEmail = await UsuarioDb.findOne({ email });
+                if (existeEmail) {
+                    return res.status(400).send({
+                        ok: false,
+                        msg: "Ya existe un usuario con ese email"
+                    });
+                }
+            }
+            campos.email = email;
+            const userUpdate = await UsuarioDb.findByIdAndUpdate(uid, campos, { new: true });
+
+            return res.status(200).send({
+                ok: true,
+                usuario: userUpdate
+            });
+        } else {
+            return res.status(400).send({
+                ok: false,
+                msg: 'El usuario no tiene permisos para acceder a esta información',
+            });
         }
-        campos.email = email;
-        const userUpdate = await UsuarioDb.findByIdAndUpdate(uid, campos, { new: true });
-
-        return res.status(200).send({
-            ok: true,
-            usuario: userUpdate
-        });
-
-
     } catch (error) {
         console.log(error);
         return res.status(500).json({
@@ -105,19 +116,25 @@ const borrarUsuario = async(req = request, res = response) => {
 
     const uid = req.params.id;
     try {
+        if (req.rol == 'Administrador') {
+            const usuario = await UsuarioDb.findById(uid);
 
-        const usuario = await UsuarioDb.findById(uid);
-
-        if (!usuario) {
-            return res.status(404).send({
-                ok: false,
-                msg: 'No existe un usuario por ese id'
-            });
+            if (!usuario) {
+                return res.status(404).send({
+                    ok: false,
+                    msg: 'No existe un usuario por ese id'
+                });
+            } else {
+                await UsuarioDb.findByIdAndDelete(uid);
+                return res.status(200).send({
+                    ok: true,
+                    msg: 'Usuario eliminado'
+                });
+            }
         } else {
-            await UsuarioDb.findByIdAndDelete(uid);
-            return res.status(200).send({
-                ok: true,
-                msg: 'Usuario eliminado'
+            return res.status(400).send({
+                ok: false,
+                msg: 'El usuario no tiene permisos para acceder a esta información',
             });
         }
